Rename common folder to types
