+++ src/context/authentication/authentication-repository.ts
@@ -1,4 +1,4 @@
-import { Result } from '../../common/Result';
+import { Result } from '../../types/Result';
 import {
   AuthenticationDao,
   AuthenticationRepository,
