+++ src/context/user/user-firebase-dao.ts
@@ -1,5 +1,5 @@
 import firebase from 'firebase';
-import { ok, Result } from '../../common/Result';
+import { ok, Result } from '../../types/Result';
 import { User, UserDao } from './types';
 
 const userFirebaseDao = (firebaseApp: firebase.app.App): UserDao => {
