+++ src/features/profile/styled.ts
@@ -1,5 +1,5 @@
 import { Typography } from 'antd';
-import styled from '../../theme';
+import styled from '../../types/styled-components';
 
 export const ProfileHeader = styled(Typography.Title)`
   margin-bottom: -8px !important;
