+++ src/types/result.ts
