+++ src/types/styled-components/index.ts
